+++ source/commands/ci/_tests/runner.test.ts
@@ -70,14 +70,14 @@ it("does not use GitHub Checks if requested not to", async () => {
 it("passes the strictm option from args into the executor config", async () => {
   const customArgs = {
     ...defaultAppArgs,
-    strict: true,
+    failOnErrors: true,
   } as SharedCLI
 
   await runRunner(customArgs, { platform, source })
 
   // Pull the executor out of the call to runDangerSubprocess
   const executor: Executor = mockRunDangerSubprocess.mock.calls[0][2]
-  expect(executor.options.strict).toEqual(true)
+  expect(executor.options.failOnErrors).toEqual(true)
 })
 
 // TODO: This occasionally fails!