+++ source/commands/utils/sharedDangerfileArgs.ts
@@ -24,7 +24,7 @@ export interface SharedCLI extends program.CommanderStatic {
   /** Use a URL to send in the DSL instead of just STDIN */
   passURLForDSL?: boolean
   /** Fail if there are fails in the danger report */
-  strict: boolean
+  failOnErrors: boolean
   /** Use GitHub Checks */
   useGithubChecks: boolean
 }
@@ -42,5 +42,5 @@ export default (command: any) =>
     .option("-c, --external-ci-provider [modulePath]", "Specify custom CI provider")
     .option("-p, --process [command]", "[dev] Runs a custom sub-process instead of the Danger JS runtime")
     .option("-u, --passURLForDSL", "[dev] Use a custom URL to send the Danger DSL into the sub-process")
-    .option("-s, --strict", "Fail if there are fails in the danger report", false)
+    .option("-f, --failOnErrors", "Fail if there are fails in the danger report", false)
     .option("--use-github-checks", "Use GitHub Checks", false)